refactor: give six oversized components a state hook beside the view
PublishDialog, CommunityGalleryTab, IsometricPreview, CutoutEditor,
BentoWorkspace and SlotConfigurator each move the prefix of their body (store
selections, local state, effects and handlers) into a sibling useX hook and
keep the guards and markup, the shape the panel sections already use. Hook
order, dependency arrays and callback bodies are unchanged.
